Admin scene creation. Interactable table view implementation
Populate admin scene with a table view. Setup table view for admin view with all relevant columns and populate it as well. Add full support for editable table only accessible to the admin. All fields are simply editable with a double click and are submitted by pressing 'Enter'. Add getAllQuestions method in the database handler to populate the admin's table view. Add get one question query from database by sending a query. For the purpose of the table view, the Question class had to be changed slightly to contain SimpleStringProperty objects instead of simple String object. Functionality stays the same with a simple .get() or .set() method. Minor: Removed check for loging in for time saving reasons. Add toString() method to Question object for debugging purposes