Thread implementation of Login and Admin page
Threads are used to perform the background tasks of querying the database in order to verify the login information of the user and for querying the database questions for the admin to see. There is a need for some kind of visual presentation of a loading for the user to understand that their login is actually going through, since now the applications just does nothing without being laggy either until the query is in fact done.