Fix remember me. Add proper account singleton
Remember me was having weird behavior since it was saving at the same time that the thread was logging in the user, but before checking if the login is indeed successful. That had as a result to save the wrong account into the file or overall have unexpected behavior since those two processes were running on separate threads. Added the functionality to the thread that handles the login now instead, so it only saves after a succesful login.
Remember me was having weird behavior since it was saving at the same time that the thread was logging in the user, but before checking if the login is indeed successful. That had as a result to save the wrong account into the file or overall have unexpected behavior since those two processes were running on separate threads. Added the functionality to the thread that handles the login now instead, so it only saves after a succesful login.